Move journal generation to src/ and generate/test as root
